fibo

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,16 +6,16 @@
  */
 int main(void)
 {
-	long long int sum;
-	long long int lastNumber = 1;
-	long long int currentNumber = 1;
+	long int sum;
+	long int lastNumber = 1;
+	long int currentNumber = 1;
 	int i;
 
 	printf("%d, ", 1);
 	for (i = 1; i <= 49; i++)
 	{
 		sum = lastNumber + currentNumber;
-		printf("%lld, ", sum);
+		printf("%ld, ", sum);
 		lastNumber = currentNumber;
 		currentNumber = sum;
 	}
